do-update: add NODOWNLOAD option

+++ do-update.sh
@@ -4,16 +4,18 @@ set -xe
 cd "${0%/*}"
 pushd "data/"
 
-for mrtfile in master4_latest.mrt master6_latest.mrt; do
-    rm -f "$mrtfile" "$mrtfile.bz2"
-    wget --no-check-certificate "https://mrt.collector.dn42/$mrtfile.bz2"
-    bunzip2 "$mrtfile.bz2"
-done
+if [[ -z "$NODOWNLOAD" ]]; then
+    for mrtfile in master4_latest.mrt master6_latest.mrt; do
+        rm -f "$mrtfile" "$mrtfile.bz2"
+        wget --no-check-certificate "https://mrt.collector.dn42/$mrtfile.bz2"
+        bunzip2 "$mrtfile.bz2"
+    done
+fi
 
 DB_OUT="dn42.db"
 
 rm -f "$DB_OUT.new"
-time python3 ../mrt2sql.py "$DB_OUT.new" master4_latest.mrt master6_latest.mrt
+time python3 ../mrt2sql.py "$DB_OUT.new" master4_latest.mrt master6_latest.mrt "$@"
 if [[ -f "$DB_OUT" ]]; then
     mv "$DB_OUT" "$DB_OUT.bak"
 fi
